backend/requirements.txt
fastapi==0.118.2
uvicorn[standard]==0.37.0
python-multipart==0.0.20
numpy==1.26.4
Pillow==11.0.0
scikit-learn==1.5.2
